Init project with basic on off neopixels
Starting point for desk_lamp.
Very simple starting point with a main.cpp where neopixels are set to blue
and we turn it on and off via a button.
No debouncing has been implemented, this is an IOT project so the button
is just a stand in for the network operations that will eventually take place.

TODO: Next step is to replace the button interface with websocket